feat: api for generate heatmap

+++ app/Http/Controllers/Users/GarbagePosts/UpdateGarbagePostController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Repositories\GarbagePostRepository;
 use App\Repositories\GarbagePostImageRepository;
+use App\Repositories\StreetRepository;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
@@ -14,14 +15,18 @@ class UpdateGarbagePostController extends Controller
 {
     protected $garbagePostRepository;
     protected $garbagePostImageRepository;
+    protected $streetRepository;
 
     public function __construct(
         GarbagePostRepository $garbagePostRepository,
-        GarbagePostImageRepository $garbagePostImageRepository
+        GarbagePostImageRepository $garbagePostImageRepository,
+        StreetRepository $streetRepository
     ) {
         $this->garbagePostRepository = $garbagePostRepository;
         $this->garbagePostImageRepository = $garbagePostImageRepository;
+        $this->streetRepository = $streetRepository;
     }
+
 
     public function update(Request $request, $garbagePostId)
     {
@@ -29,12 +34,32 @@ class UpdateGarbagePostController extends Controller
 
         try {
             DB::beginTransaction();
-            $garbagePost = $this->garbagePostRepository->findOrFail($garbagePostId);
+            $request->validate([
+                'description' => 'required',
+                'street_id' => 'required|integer',
+                'date' => 'required|date',
+                'before_images.*' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048', 
+                'after_images.*' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048', 
+            ]);
+
+            $garbagePost = $this->garbagePostRepository->find($garbagePostId);
+
+            if (!$garbagePost) {
+                return response()->json([
+                    'message' => 'Post does not exist',
+                ], 400);
+            }
 
             if ($garbagePost->user_id !== $user->id) {
                 return response()->json([
                     'message' => 'You do not have permission to update this post',
                 ], 403);
+            }
+
+            if (!$this->streetRepository->find($request->street_id)) {
+                return response()->json([
+                    'message' => 'Street does not exist',
+                ], 400);
             }
 
             $this->garbagePostImageRepository->deleteByCondition([
@@ -43,8 +68,9 @@ class UpdateGarbagePostController extends Controller
 
             $garbagePostData = [
                 'description' => $request->description,
-                'locationable_type' => $request->locationable_type,
-                'locationable_id' => $request->locationable_id,
+                'street_id' => $request->street_id,
+                'latitude' => $request->latitude,
+                'longitude' => $request->longitude,
                 'date' => $request->date,
             ];
